Avoid defining baseUrl in tsconfig
Originally baseUrl was set in order to be able to import `render` in
test files without relative paths.
Anyway, because it was defined in project-wide tsconfig it affected
the whole project and caused unexpected errors in development:
e.g. creating `src/leaflet.ts` file in project caused
`import * as L from 'leaflet';` style imports to break, as TS tried
to find those from our local file.

Technically we could define baseUrl in `tsconfig.test.json`, but then
doing project-wide TS type checking would have to be done in two steps
(one for test files, one for everything else) instead of all at once
(as we are doing right now). So for the sake of simplicity it might
be better to do this change now and consider other options if relative
imports become too annoying.