Use `react-hook-form` + `zod` instead for `formik` + `yup`
**Formik vs. react-hook-form:**
I did a little bit of digging and current public opinion seems to be that
react-hook-form is just better than formik. See e.g. following
blog posts:
- https://www.framelessgrid.com/react-hook-form-vs-formik-react-form-library-comparison-in-2021/
- https://blog.bitsrc.io/react-hook-form-vs-formik-form-builder-library-for-react-23ed559fdae
- https://blog.logrocket.com/react-hook-form-vs-formik-comparison/

**yup vs. zod:**
Inferring TS types from yup schema works in theory, but in practice
I started already facing problems with it even in this project.
In other hand inferring types from zod schema seems to just work,
at least so far.
In context of form validation yup's default validation messages
seemed to be more ui friendly, but at some point we'll have to
customise those anyway so that is something that we can live with.
And luckily react-hook-form supports both (and plenty of other
validation libraries too) so if problems arise we still have
options to change direction.