Fix route links and stops being saved in wrong order
Until now, the route links and stops were accidentally saved in the
order they were returned in by a query containing an IN clause. The
order was in fact arbitrary.

This commit fixes this by retrieving all links and stops in the same
graphql query and re-ordering them according to the route links returned
by the map-matching backend.

Additionally, the graphql types used in the query were corrected.