Refactor how stops are handled in UI
Original implementation of stops handling was implemented back in
the days when we didn't yet have backend instance.
Because of that stops were handled 100% in UI.
At some point the implementation was changed so that we fetched stops
to the UI from the backend, but then kind of combined stops from
backend with stops in the UI and the result was really confusing.

After this refactoring things are still quite messy, but perhaps still
better than they were.

- Now it is possible to delete stops from the backend via the UI
- Now we inform users about the fact that dragging stops isn't actually
  supported right now

There are still major issues:
- When new stop is created, it is not fetched to the UI (and temporary,
  corresponding unsaved stop in the UI won't be removed)
- Editing existing stops won't work and actually just results duplicating
  of existing stop (!)