Add CI job for e2e testing using a specific branch
The script looks for a branch with the same name in the jore4-robot repository. If it exists, it runs the tests from that branch. If not, it'll return with success.

This is useful when developing the UI and it required changes also in the e2e tests. Just create a branch with the same name in the jore4-robot repository and fix the tests there.