Simplify map style file
This removes following warning from console:

```
Vector tile source "parkandride" layer "facilities" does not use vector tile spec v2 and therefore may have some rendering errors.
```

As a drawback the style file is not anymore 1:1 to the original hsl
style file, but that's probably not an issue as there were no specific
reasons to use the whole file other than that it worked out of the box.

Also `react-map-gl` v7 had this same warning and this commit fixes also it.