Improve the create route e2e test reliability
Previously the test did not wait for the route properties form to close after saving, so sometimes the clicks meant for the map were actually registered for the form.
Also now clicking way before the first stop and way after the last stop when drafting the route to make sure that the snapping handles are not overlapping with the stop markers.