Prettify generated graphql types
As we start using the generated custom graphql types more and more, it's important that they are human readable. Thus using prettify on this file too.

I tried using prettify as a graphql-codegen hook as suggested in its documentation, but it does not always get called. Thus calling prettier directly in the package.json script