Fix docker build
Unluckily we have to include also test-db-manager to docker build as
it is now listed as a devDependency on ui's package.json file.
Otherwise docker builds would fail.

Maybe we could avoid that by re-organizing ui's package.json file
so that all the dependencies required for _builds_ would be in regular
dependencies and rest of the dependencies would be on devDependencies,
as then we could install only normal dependencies on docker context
and get rid of the original error.