Add support for all-option to VehicleModeDropdown
Currently the dropdowns are only used when creating a line or route and they are required fields,
therefore they do not have the 'All' option. But when we use them as search filters, we need the
'All' option. To reuse these same components, we add the support for 'All' option as well. This
commit will add the support to EnumDropdown and VehicleModeDropdown.