Change obserationDate query param init to optional
Before this the observationDate was initialized automatically to query parameters
if the component uses the useObservationDateQueryParam hook. The initialization
is unwanted if the component only wants to read the observationDate, and the value
is optional. For example in this case the useShowRoutesOnModal hook only uses the
observationDate if it is already there. But we do not want to initialize it if it is not there.

Before this we used useShowRoutesOnModal hook with every route and line row, which
ment that the observationDate query parameter was initialized n times, where n is the
amount of rows rendered..