Use "real" data in vehicle service table
We currently have limited seed dataset and limited ui that doesn't allow
user to select line for timetables, so  in this commit we show all
data from db as an example.

In near future we should improve seed dataset and only show data based
on selected line id in this ui.

Related to HSLdevcom/jore4#950