Use component specific attribute fetching for table rows
We fetched ...route_all_fields and ...line_all_fields for component that does not
need all the fields. It is better to use the component specific fragments. This will also
make it easier to re-use this component when we specify only the fields that are needed. Otherwise
we would have to fetch all the fields every time.