Fix route creation with open validity end
When setting the validity end for a route to "open end", it is set to
'null'. This was accidentally caught by a check, which was supposed to
verify that the route data was set at all, because it tested for
falsiness instead of 'undefined'.

This commit fixes the check to explicitly test for '=== undefined'.