Add timetables search and refactor logic to support generic components
- We use the same components for timetables- and routes and lines -search. The functionality will
be different depending on in which path we are ('/timetables' or '/routes').
- Change 'Path.lineTimetables' from '/lines/:id/timetables' to '/timetables/lines/:id'
- It is now RouteLineTableRow's responsibility to check if the basePath is 'timetable'. If it is
anything else, we render RoutesAndLines version of the line. This means that the main functionality
(click the row) will take the user to line details and the row will render button to tak the user
to line timetables.