Adjust stop filters on map according to stop changes
Resolves HSLdevcom/jore4#879
When editing or creating a new stop, we make sure that after the edit / creation the stop won't
dissapear due to map stop filtering. So if we for example create a stop with draft priority and the
drafts are not currently shown on map -> change the filters so that drafts are also shown on map.