Add waits to timingplace dropdown search
We need to type in the timing place name to be sure to find it in the test, but without the wait
cypress picks up a "deprecating" html element where it does not find the result. When we add the
wait, we should have the correct element with the result before the
`.find('li').contains(timingplace)` is ran. Same for the assert in the last test.