Fix Apollo dev tools extension
Was happening because next.js renders our app (ui/_app.tsx) twice,
and the cache is initialized during that process.
This seems to be because of next.js rewrites, and is normal behavior:
https://github.com/vercel/next.js/issues/33028#issuecomment-1006179699

However, the Apollo dev tools extension doesn't work properly (shows nothing)
when multiple apollo clients are created. See:
https://github.com/apollographql/apollo-client-devtools/issues/822#issuecomment-1059166308