Add new fetching logic to timetable page
This commit is pretty massive, but the roots of the refactoring and the new logic were deep. So this
commit replaces the old way of fetching vehicle schedules for vehicleScheduleDetails and also adds support
for substitute operating days by line type. Earlier we had this a bit awkward logic of 'fetch all data in two separate
requests and then combine the information to compile wanted end result and this resulted in only the in effect schedule
even though this style fetched a lot of redundant data.
So with this new logic we get all the necessary information from database function and then we play with it in different
views. Now we get all vehicle schedules for day types on the observation date, we combine them in to vehicleJourneyGroups
and calculate which is the one in effect.

Resolves HSLdevcom/jore4#1158