Cancel move action if the stop move is prevented by hastus options
When a stop is being moved, the move action might be prevented if the stop is used as
hastus place in some route. After displaying the error message, the stop move was still
enabled, even though it is now clear that the stop cannot be moved before actions have
been taken. So in this case we should disable the stopmode.

Resolves HSLdevcom/jore4#1182