Rename routes to reducedRoutes to be more precise
This reduced set has 1 direction per route label. The result set is initially
reduced so that we do not show 2 rows per route label (1 for each direction).
The renaming is now more clear that it does not have all the routes that we get
from the result.
Also this is because in the next commits we also add the whole result set and use it
and for that we want to use 'routes' as the variable name.