Refactor selecting and exporting routes and lines
The initial logic had a little bit too much responsibility, and we saved to state what were the
selected exported routes, even if we selected a line. So the flow was that if you select a line,
it actually saved all the routes within that line to be "selectedRouteLabels". And also to check if
a checkbox should be checked visually, we had to see if the selected routes matched the line's
routes, which is unnecessarily complicated.
This logic is now refactored to be simpler: you select rows, doesn't matter if it is route or a line.
"selectedRows" will always have the selected rows label (uniqueLabel in case of a route).
And then when we export, we go through what should actually be exported.
Also removed unintuitive behaviour, where all of the routes were exported when none
was selected.