Reset export selections on search
Initially it was possible to do multiple searches and select rows for
export from different searches. This was not intentional. If we want
this kind of feature in the future, we would also need to indicate with some
kind of preview of selectations, what is actually selected. But for now,
we just reset the selectedRows on search.