Make timingsettings on route dependant on each other
A stop has to be set as timing point in order to use it as regulated
timing point. Also stop has to be regulated timing point in order to
allow loading time. So we clear the value and disable the checkbox if the
value it depends on is set to false.

Resolves HSLdevcom/jore4#1102