Refactored RouteStopsHeaderRow into RouteRow
The old RouteStopsHeaderRow was confusing, because it was not a header row per say, nor does it contain Stops.
It is just an expandable row with route information (=>RouteRow).

Along this refactor, changed the names related to the expansion functionality to be "Expand" related, instead of "Open", for claritys sake.