Add component test for PassingTimesByStopTable
There is a few functions here that are used only to ease out creating
test data. I'm not sure if that is the best way to do it, but without
them initializing test data would clutter up the "whole" file.

Data used in tests shows that there is lot of unused data. That
should be refactored to only take in as props what it really needs.

Resolves https://github.com/HSLdevcom/jore4/issues/1265