Update pinned hasura image
As explained in the committed comment, we removed the FK constraints from return_value.timetable_version
because we do not want to make it part of the data model, since it is only a dummy table for function return value.
But because of this, the graphql-codegen now types it as nullable, even though it can't ever be null in our case.
To fix this, added an if-statement with an explanatory comment.