Add route export to timetable import tests
Route export is required to be done before timetables are imported for that route

There was a conflict between the `usedAsTimingPoint` flag of the 2nd stop
point of the journey pattern and the Hastus import file.

Add one infrastructure link and stop point more for properly testing handling of middle stop points.

Resolves HSLdevcom/jore4#1494