Remove the unused showLabel parameter from PriorityForm
This was added for timetables form case,
which is in a separate component now.
No point keeping unused code.

Also thought about removing the hiddenPriorities parameter,
since it is only called with same values at the moment.
Decided to keep it though, since current usage
is a special case for lines, and this component could be used
eg. for routes as well.