Revert removing rewrites -function
As we now know, this is actually used for the client side routing. Bringing this back will cause a warning to appear
when running the frontend: `Specified "rewrites" will not automatically work with "output: export".
See more info here: https://nextjs.org/docs/messages/export-no-custom-routes`. This warning should be irrelevant for us.