Add protected routes
Opening the site or refreshing a page will trigger userInfo fetching and if the user is not logged in, redirect
to front page. This is a kind of quick fix for the test users and this logic might change once we get jwtTokens and properly
user roles.

Also move visit to beforeEach so that the mockLogin works properly. Now the tests
are slow in dev environment, but they are not dependant on each other anymore.