Add graphql unauthorized temporary error handling
This handling is a quality of life change for developers. If
you stop and start depdencies, it will obviously remove and start auth
service, and you are no longer authed, but if your site is still open and
the frontend still thinks you are logged in but all graphQL queries return
unahtorized. This change will reload the page -> see that you are actually
not logged in -> redirect to frontpage and inform the user to log in.

We probably should and will remove this behaviour later when we get different
user access roles.

This and commit: faf669ee56982ab6e87c8899bf16e6bf47625dff combined
Resolves HSLdevcom/jore4#1291