Change e2e gql search waits to check for loading indicator
Before this change the only way to check if the loading was complete, was to wait for the
gql queries to complete, which is not of course optimal for e2e tests, because the user
can't see the requests. But now we have loading indicator so we can do the check properly.