Add dialog error modal to timetables import page
The previous toast style of error messages was not good enough,
especially now that the Hastus importer returns a better, and longer,
description of the actual error.

Followed the design of the retryable popup, but made it take in elements
to get a versatile error dialog card. This error can't be fixed
with a retry, so it doesn't have the retry function.

Also adds an `ErrorDialogItem`, which is made for the Hastus import errors,
but could most likely be used elsewhere with other errors.