Fix run-integration-tests actions order
setup-node action is run after we install dependencies, which can cause the version restriction
to fail to incompatible versions if the cache has another version at that time. So we need to first setup the
correct node version, and then run yarn install. This is also the order in all other workflows.