Fix stop order rendering
This rendering relied that passingTimesByStop was populated in the correct order. It seemed that
there was a possibility that this was populated partly wrong, or it used partly some cache(?) or it was not
immutable or something, because we had cases where the first stop was actually rendered last and the
only possibiity is that the last stop was added last to passingTimesByStop object, which will put it
last in the array when we use Object.entries(passingTimesByStop).

The mapping logic is now changed so that we always map stopsInJourneyPattern, which always should
have the stops in correct order, and changes to timetables won't change that. It also is a bit more readable
to have the mapping this way.

Also required fix to `<PassingTimesByStopTable /> should render labels and times correctly` test
which was missing the H221 stop from the route journey pattern.

Resolves HSLdevcom/jore4#1547