Add ordered_ prefix to journey_pattern_with_stops fragments attribute
We have this fragment, which orders the stop points by scheduled_stop_point_sequence: asc,
but this is not indicated in any way, so by adding this prefix we make clear in the code that this particular
fragment will have it's contents ordered. Otherwise there is no distinction between this and some other
query which does not order the stop points.