Update seed data dump for network and routes.
The new dump contains a fix to the route names, which were previously the same
between the inbound and outbound directions.

The new dump also has a lot of new lines and routes because the jore3-importer
logic has been improved by resolving conflicting time periods between lines and
routes.

Resolves https://github.com/HSLdevcom/jore4/issues/1651