Fix an e2e failure due to year changing
The TimetableVersionTableRow component shows a status text
for each vehicle schedule frame, which is "Voimassa" for a frame that is
still valid today and priority otherwise.
The frame validity ended at the end of last year,
aka not valid anymore, thus the text changed and tests failed.

Could maybe fix by mocking the current date in browser/app or something,
but don't see this as worth the effort so just kicking the can down the road (hi 50s?).