Add titles to buttons & icons in search result rows
Also replace custom `Tooltip` with browsers built-in title.
The Tooltip was unnecessarily difficult from an accessibility point of view, even though it looked and worked great visually.

This commit has some VERY DRAFT changes, I hope someone could comment on the `RouteLineTableRow` component and it's generic way to handle the type of the Row (Lines&Routes vs. Timetables).It causes complexity with my implementation (`VehicleIcon`) of figuring out whether it's a Line, Route or a Timetable row, which needs to be known to get a human readable title out. It started to feel like it's running out of hand with the translations and figuring out the types, so it needs someone elses opinion...