Add spinner to each route row on line details page
The line details in general load fast enough that it wasn't
deemed necessary to have a whole page spinner.
The route details on the other hand might take a moment
(less than a second really, on dev), so show a loader animation until each is ready.

Hardcoded the row height to help ensure the loader height matches actual row
and to make it more robust against small layout changes.