Fix type of the stop parameter in StopPopup
The fragment was only used to generate the TS type for the parameter.
But nothing was enforcing that the stop actually contained the properties
in the fragment.
Actually the fragment was out of date and contains a lot more properties.
Kept the type casting for simplicity, but use the type from original query instead.