Refactor VehicleJourneyGroupInfo parameters to require only necessary
This component had unnecessary requirements for parameters. It required `VehicleJourneyGroup`
as parameter, even though it did not need all attributes within that type.
Also the VehicleJourneyGroup made the requirements to be so that vehicleJourney's had to be
type `VehicleJourneyWithService` which reduced the reusability greatly or forced to use that fragment.
Only the start_time's are required for this component which is now the only required attribute. (<Pick..>)