Add timetable version details panel functionality
This version details will open the timetable schedules of the selected version, which corresponds to the selected
versions vehicleScheduleFrame's schedules. This view is now currently the only way you can view all the timetables
within one vehileScheduleFrame, which is crucial if the user wants to change the validity of a timetable, because
it affects the validity period of the whole vehicle schedule frame.

This is pretty big commit and it has some overlapping components with existing code, and the existing code can be also refactored
to use some of the new components. For example `TimetableHeading` can be relocated and used with the existing

* Added a optional `onChange` callback function to `ChangeTimetablesValidityModal`, so that we add custom
functionality in case of change e.g. re-fetching some view data.

Resolves HSLdevcom/jore4#1577