Add darker version of tweaked brand blue
To be used as hover color, for example.

Can't mutate this easily from existing colors with current tools.
Would require converting HEX -> HSL -> adjust the L -> back to HEX,
which can't be done with plain CSS or Tailwind.
Since this is currently a one off, didn't want to add a JS color library,
but should consider this if there will be more similar situations.

Also noticed the relation between brandDarker and tweakedBrand colors,
so added a comment about that.