Add e2e tests for TimetableVersionDetailsPanel
Test that the panel opens and show's the correct information. Also test that
the validity period can be changed and updates correctly.
Add all the necessary testIds. Also some minor changes e.g. change inline
testId in to testIds object.
Also added some folder structure to the pageObjects where we can start to move
some of the pageObject files.
As last note: there are still guidelines open for e2e testing and how we want to implement
and use these pageObjects, also how we use the base data for each test. Currently we have
just all of the base data copy-pasted in each test, which is not so nice. We should probably
create some base datasets that can be used in multiple tests, and if there are some spesific
data that needs to be inserted in a new test, we could either do the modification on the test or
just expand the base dataset.