Add LoadingWrapper to CommonSubstitutePeriodForm
This is just a split commit from ffb16f2cc6d6b067a1c75e87136c9734767adb24 [ffb16f2c] which had
multiple changes in same commit. Have to split these to understand what is happening. This loadingWrapper should
probably be in CommonSubstitutePeriodSection rather than in the form, but that is for later.