Add e2e tests for common substitute days
This is minorly fixed and modified commit from an earlier implementation
(done in this branch), not everything is perfect even now, but the tests
work. This might need some adjustment and more cleaning in the future.
There is still the minor possibility that
there are "old" toasts visible when the assertion is done, so we might
need to handle that. Best solution would be to add a possibility to close
a toast message in UI, and then in these tests always close the toast after checking.