Fix e2e occasional substitute day test
We saved an occasional substitute day "in the future" so that the observation period should change due to that.
We made an assertion that the observation period is changed and that we get a toast message for it. But since the test
did not explicitly set the observation date and just used the default period, which is 1 year from "today". And since the
occasional date was saved as 8.3.2025 this date would be inside the observation period when we hit 8.3.2024. That day was today
and the test broke. So by explicitly setting the observation period, this test should always work.