Fix editRouteShape basedata
Stops on the base data are in wrong order and placed in wrong infralinks. The existing basedata
makes the route go in wrong direction causing it to make u-turns. Also the stop that is located
geographically on infralink 1 is linked to infralink 3. Same is with stop that is located geographically
on infralink 3, is linked to infralink 1.
NOTE: this is just iteration, and the data will be unified between the tests and moved to datasets.