Add stopPlace utils and unit tests
Tiamat has some lazy data structures, which uses arrays to save multiple different data, e.g. `keyValue` and
`alternativeNames`. Now if we want to update some precise data inside those arrays we need the initial data
as well as the new data, in order to update or insert the data in the array without losing data. This is where the
util functions come in to play. You can now give the initial array and the new values, and they will be upserted.

Also added couple of unit tests for these util functions.