Add StopRegistryTransportModeTypeSubset enum
The enum is generated from the tiamat database, which also has types like 'Air', 'Funicular' and 'Cableway'
which we are not using. This subset enum is created, so that we don't need to handle the options that are not used.
Currently it will be only used temporarily for EnumDropdown component, which creates the options from all the enum values.
But this is just temporary untill tiamat supports multiple transportModes for a stop place.