Add StopPlace basic details edit functionality
Add the edit mode to basic details. There are couple of helper components
in the forms (form rows) to help with the form boiler, but the test id's
use the parent's testId prefix.

Other things to note: we can't achieve any kind of transactionality in frontend for the requests,
and we are heavily trusting that the second update (to tiamat) goes in successfully if the update
to routes and lines database is successful. But there is already at least one case where this is not
the case: tiamat has e.g. an constraint that a 'rail replacement' can only be applied to bus typed stop.
This is something we need to restrict in front end and that is not yet done in this iteration.

Minor things: there are some TODO's still present, some are waiting for tiamat changes, some are
waiting for the next iterations of this edit (e.g. editing stop label)

Resolves HSLdevcom/jore4#1680