Add StopPlace basic information edit functionality
Now this is a big one and some parts are still under consideration. So here
we add the edit mode to basic information. The form files are pretty simple, but
it is still under consideration that should all the form fields still be in only one file,
or possibly have the form fields in their own components, but have the schema
in the "parent" file.

Other things to note: we can't achieve any kind of transactionality in frontend for the requests,
and we are heavily trusting that the second update (to tiamat) goes in successfully if the update
to routes and lines database is successful. But there is already at least one case where this is not
the case: tiamat has e.g. an constraint that a 'rail replacement' can only be applied to bus typed stop.
This is something we need to restrict in front end and that is not yet done in this iteration.

Minor things: there are some TODO's still present, some are waiting for tiamat changes, some are
just notes and will be removed in the final version of the PR. As is this commit message.

Resolves HSLdevcom/jore4#1680