Upgrade apollo client and apply remove typename
As described in https://www.apollographql.com/docs/react/api/link/apollo-link-remove-typename/
if we use gql responses as base for further mutations, the apollo's included __typename will cause
problems. By adding this removeTypenameFromVariables to our apollo link, the __typename's will
be removed from operations automatically and we don't need to do that manually. This feature is
available from apollo client 3.8 onwards, but with the same pace upgraded to latest (3.9.10)