Display a warning if stop main line selection and related signage don't match
That is, if main line has no main line sign
or main line sign is set despite stop not being a main line stop.

To keep things simple(ish), decided to use a text that kind of works for all cases.
More specific messages would depend both on the error case
and on which section the warning is shown,
so didn't see further detailing worth the added complexity.