Refactor handling of preventing stop place mutation from resetting the description and add E2E tests
To centralize handling of this and remove some copypaste.

Would be even better if this could be enforced on the mutation input type
level but couldn't figure out a way to do that (easily at least).

At some point we might have to add similar handling to other tables
that this mutation could modify,
but those might be better handled with separate helpers.