Refactor stop search query logic to target tiamat
Our stop search was initially created to target routes and lines database and then
we used the remote schema relationship to also fetch data from tiamat database.
Majority of the upcoming search criteria are attributes are that are located in the tiamat
database and we want to use the query level filtering (where-condition) as much as we can.
We can only use query level filtering to the fields that are located in the database which the
query is targeting. In other words, if we target the query on routes and lines database's
scheduled_stop_point_id, we can only use query level filtering for label, validity and priority.
But all the other criteria would have to be done on frontend. These are all the name variations,
stop states and so on. And that is why we want to target tiamat database with this search query.

So this commit changes the query main target so that we primarly fetch and filter data from tiamat
and then we enrich the result with the routes and lines database data.

Resolves https://dev.azure.com/hslfi/JORE%204.0/_workitems/edit/35447