Add a new e2e task for fetching infrastructurelink ids by external ids
Currently we are fetching the infralink ids by external ids just using a fetch,
but the problem is that the resulted infralink ids are not in the same order than
the external ids were. Yet, they are treated as they would have been in the same order.
This results in stops and routes being linked to wrong infrastructure links.

By creating a new task, which handles the mapping and ordering the result in the same
order as the external links are, we reduce boiler on the tests.