Add e2e base dataset and use it on editRouteShape
To reduce boiler plate on each test file, where we create the same routes over
and over again, we now have a base data set that can be taken in to use in multiple tests.
This dataset also will have a bit more data in it, so that we can test features a bit more
thoroughly.

Also refactor editRouteShape and some page objects.
- Copied the setPriority functions from PriorityForm to ChangeValidityForm,
since the PriorityForm could not be used if there were more than one
priorityForms on the screen. For now, I did not move it completely
(so it still is in the priorityForm), thats maybe for later if needed.
- Added position to MoveRouteEditHandleInfo, but take a note that this will
probably be gone anyway soon after mapGL upgrade.
- Removed createRoute function usage from editRouteShape, that masks
way too much test logic, also it is not usable for very many situations.
- Still using fillRouteProperties for the text fields here. It could also
benefit from an refactor iteration round, because it has way too much
conditional test logic which should not be used. But thats for later.
- Now the test uses loading indicators instead of listening to network requests.
- NOTE: we could/should probably disable save buttons during loading, because
it is possible to save things before the newest changes has been applied. It is
probably not a real problem, but in theory.