Add base dataset and use it on editRouteShape test
Okay, so to reduce boiler plate on each test file, where we create the same routes over
and over again, we now have a base data set that can be taken in to use in multiple tests.
This dataset also will have a bit more data in it, so that we can test features a bit more
thoroughly.