Style route drawing and limit draw modes to 1
When drawing we only want one mode and that is drawing. All other functionalities are handled by our custom implementation. Because of custom implementation, we need to manually fix the order of the layers when drawing a route. This is done by moving layers that come after draw layer to below it.

Note: MapBoxDraw adds suffix `.cold` to layer id.