Add base dataset and refactor e2e test hastusExport
While adding the base dataset had to do a pretty big refactor at the same time.
- Currently the 'should export a line' is skipped, because it brings false sense of security
since it practically is the same test as 'should export a route'. We need to add more routes
to the exported line before removing the skip, so that it actually tests something.
- Removed some test antipatterns. Now the tests modify the timing settings via UI instead of
different test setups and helper functions which greatly reduced the readability of the tests.