Add direction to ExpandableRouteRow pageObject
Add the precision to find the right row with the added direction part on testId.
Each of the route rows fetch their own information, so they are not ready at the same time. That is why
it would be flaky to use for example "getExpandableRouteRow().first()" . The other solution would be to wait until
no RowLoaders are visible and then just use the route label, but for this iteration this is good enough.
Also fixed toggleRouteSection to use "findByTestId" instead of "getByTestId" which starts a new chain.