Remove routeShouldBeSelected pageObject assert
This asserting method did practically nothing but mislead the reader. `this.getHeader().get()` is the same as
`cy.get()`, because the chained get just starts a new chain. So this function actually just checked if it could find
div that has `routeName` anywhere in the dom.

For this iteration I did a minor replacement function that gets the header with the given label and direction. This
probably will not be the final best solution, but for now it does things more correct, and also we get the direction
check here aswell.