Refactor PassingTimesByStopTable e2e test and use base dataset
- Added 'data-highlighted' to PassingTimesByStopTableRowPassingMinute to make the
highlight checks more reliable than checking
- Refactor the test to use new pageObjects, because the old implementation did not
work with the new wider base data, and it would have required modifications to the old
methods, but the old method (classes with constructors) were a bit too heavy and
complicated for this purpose. It is better to just use pretty simple page objects and then
just use .within() in the tests. It makes the tests more debuggable and readable.
- Also added another test which checks the arrival times - the old test said that it would
check the arrival times but actually there were no checks at all.