Use the accessibility level stored in Tiamat when displaying in UI
To hopefully catch possible bugs with this a bit more easily.
Otherwise it would be hard to notice if we persist an incorrect
value to Tiamat, since it was shown nowhere in the UI.

The GraphQL asserts in E2E tests are not stricly necessary anymore
now that the value also affects UI, so removed those.