Fix the issue where you can't go back from the search results page
After an initial search, you were stuck on the stop results page. Trying to navigate back to the initial search page would trigger a new search and went back to the search results page. This was caused by the old `useEffect` code. Now it will not trigger a search on an initial loading.