Add observationDate to stopDetails and show highest priority
To show the highest priority stop details on given observation date we also had to do couple of
other changes. Mainly the URL is now `/stops/:label` instead of `/stops/:id` so that we can use the
label and observation date to show the correct details. Also updated the stopDetails links to use the label
instead of id.

There is also minor changes to StopTitleRow etc. so that it accepts null and undefined values when the
observationDate + label doesn't find any results (stop is not valid on chosen date)