Fix observationDate gql filters on stop details
We checked that the observationDate had to explicitly be before the
stops validity_end, and explicitly be after the stops validity_start. But missed
that the validity_end and validity_start are also nullable fields. So it is possible
that a stop does not have e.g. validity_end set, which would not have taken in to
result set without these changes. Now we include the ones that doesn't have validity_end
and/or validity_start set.
NOTE: The UI forces the user to set validity_start, but despite that the field in database is
nullable and that is why we have to deal with that one as well.